Добавил бафы для противников (хп, скорость, урон)

+++ Assets/Scripts/Scenes/GameScene/Contexts/GameContext/Abstracts/Interfaces/IRegenerationSpeedBuff.cs
@@ -4,7 +4,7 @@ namespace GameContext.Abstracts.Interfaces
 {
     interface IRegenerationSpeedBuff : IBuff
     {
-        public void IncreaseRegenerationSpeed(int value);
-        public void DecreaseRegenerationSpeed(int value);
+        public void IncreaseRegenerationSpeed(float value);
+        public void DecreaseRegenerationSpeed(float value);
     }
 }
